==> Final Codes/04. DataSplit.ipynb <==
# Importing the dataset

import pandas as pd

from google.colab import drive
drive.mount('/content/drive')

HotelReview = pd.read_csv("/content/drive/MyDrive/BAP Project/Datasets/Final Datasets/Hotel Reviews Enhanced Dataset (Latest).csv")
HotelReview = HotelReview.drop(["Unnamed: 0"], axis=1)
HotelReview['Year'] = HotelReview['Dates'].str.strip().str[-4:]
HotelReview = HotelReview[HotelReview['Year'].astype(float) >= 2007]

print (len(HotelReview.index))
print(HotelReview.info())
print(HotelReview.head())


for col in HotelReview.columns.drop(['Titles','Reviews','Dates','No. of Reviews','Address']):
    print('-' * 40 + col + '-' * 40 )
    display(HotelReview[col].value_counts(dropna=False).sort_index(ascending=True).head(10))


#Data Exploratory Analysis

#check for rename/ NA/ impute, datatype, polarity score

#check for NA values
print (HotelReview.isna().sum())

#rename col names
print(HotelReview.columns)
HotelReview.rename(columns = {"Review.Rating": "Review_Rating", "Travel.Type": "Travel_Type", "No. of Reviews": "No_of_Reviews", "Hotel Class": "Hotel_Class"}, inplace = True)
print("\n After modifying columns:\n", HotelReview.columns)

#impute TravelType 'others'
HotelReview['Travel_Type'].fillna("Others", inplace=True)
print('\n Dinsinct values for TravelType:\n' + str(HotelReview['Travel_Type'].unique()))

#NA drop rows (Titles, No. of Reviews, Rating, FreeParking.. RoomService)
HotelReview = HotelReview.dropna()

# Data types

print(HotelReview.dtypes)

HotelReview[['Hotel','Titles', 'Reviews', 'Address']] = HotelReview[['Hotel','Titles', 'Reviews', 'Address']].astype("string")
HotelReview[['Review_Rating','Travel_Type', 'Hotel_Class', 'Rating','Location','Cleanliness','Service','Value','FreeParking','FreeWifi','FreeBreakfast','Gym','Pool','Restaurant','BarLounge','TV','EventFacilities','RoomService']] = HotelReview[['Review_Rating','Travel_Type', 'Hotel_Class', 'Rating','Location','Cleanliness','Service','Value','FreeParking','FreeWifi','FreeBreakfast','Gym','Pool','Restaurant','BarLounge','TV','EventFacilities','RoomService']].astype("category")
HotelReview[['No_of_Reviews', 'Year']] = HotelReview[['No_of_Reviews', 'Year']].astype("int")
print(HotelReview.dtypes)


#Add polarity scores
import nltk
nltk.download('vader_lexicon')
from nltk.sentiment import SentimentIntensityAnalyzer

sia = SentimentIntensityAnalyzer()
positive = []
negative = []
neutral = []
compound = []

for rev in HotelReview['Reviews']:
  score = sia.polarity_scores(rev)
  positive.append(score['pos'])
  neutral.append(score['neu'])
  negative.append(score['neg'])
  compound.append(score['compound'])

word_count = []
for sent in HotelReview['Reviews']:
  count = sent.count(' ') + 1
  word_count.append(count)

HotelReview['Reviews_WordCount'] = word_count
HotelReview['Sentiment_Pos'] = positive
HotelReview['Sentiment_Neg'] = negative
HotelReview['Sentiment_Neu'] = neutral
HotelReview['Sentiment_Compound'] = compound

HotelReview['Sentiment'] = HotelReview['Sentiment_Compound'].apply(
    lambda x: 'Positive' if x >0 else 'Neutral' if x==0 else 'Negative')

HotelReview.info()

HotelReview.head(3)

from sklearn.model_selection import train_test_split

train_set, test_set = train_test_split(HotelReview, test_size=0.2, random_state=42)

HotelReview.to_csv('/content/drive/MyDrive/BAP Project/Datasets/Final Datasets/Hotel_Cleaned.csv', index=False)
train_set.to_csv('/content/drive/MyDrive/BAP Project/Datasets/Final Datasets/train_set.csv', index=False)
test_set.to_csv('/content/drive/MyDrive/BAP Project/Datasets/Final Datasets/test_set.csv', index=False)
